requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Jinja2==3.0.1
itsdangerous==2.0.1
click==8.0.1
MarkupSafe==2.0.1
openai==0.27.0
python-dotenv==0.19.0
flask-cors==3.0.10
gunicorn==20.1.0
transformers==4.30.2
torch==2.0.1
nltk==3.8.1
scikit-learn==1.2.2
pandas==2.0.3
numpy==1.24.3
spacy==3.5.3
python-Levenshtein==0.21.1
requests==2.31.0 